requirements.txt
python-telegram-bot==20.7
pandas==2.2.2
python-dotenv==1.0.1
apscheduler==3.10.4